fix2

+++ modules/long-arithmetic/src/long_arithmetic.cpp
@@ -17,7 +17,7 @@ bigInt::bigInt(std::string str) {
             this->negFlag = false;
         }
 
-        for (long long i = str.length(); i > 0; i -= 9) {
+        for (__int64 i = str.length(); i > 0; i -= 9) {
             if (i < 9)
                 this->storage.push_back(atoi(str.substr(0, i).c_str()));
             else
@@ -44,7 +44,7 @@ std::ostream& operator <<(std::ostream& os, const bigInt& bi) {
         if (bi.negFlag) os << '-';
         os << bi.storage.back();
         char old_fill = os.fill('0');
-        for (long long i = static_cast<long long>(bi.storage.size())
+        for (__int64 i = static_cast<__int64>(bi.storage.size())
             - 2; i >= 0; --i) os << std::setw(9) << bi.storage[i];
         os.fill(old_fill);
     }
@@ -97,7 +97,7 @@ bool operator <(const bigInt& left, const bigInt& right) {
             if (left.storage.size() != right.storage.size()) {
                 return left.storage.size() < right.storage.size();
             } else {
-                for (long long i = left.storage.size() - 1; i >= 0; --i) {
+                for (__int64 i = left.storage.size() - 1; i >= 0; --i) {
                     if (left.storage[i] != right.storage[i]) return
                         left.storage[i] < right.storage[i];
                 }
@@ -227,7 +227,7 @@ const bigInt operator /(const bigInt& left, const bigInt& right) {
     b.negFlag = false;
     bigInt result, current;
     result.storage.resize(left.storage.size());
-    for (long long i = static_cast<long long>(left.storage.size())
+    for (__int64 i = static_cast<__int64>(left.storage.size())
         - 1; i >= 0; --i) {
         current.shiftRight();
         current.storage[0] = left.storage[i];
